Add an abstract query class
