+++ snuba/query/processors/mandatory_condition_applier.py
@@ -13,7 +13,7 @@ class MandatoryConditionApplier(QueryProcessor):
 
     def process_query(self, query: Query, request_settings: RequestSettings) -> None:
 
-        mandatory_conditions = query.get_data_source().get_mandatory_conditions()
+        mandatory_conditions = query.get_from_clause().get_mandatory_conditions()
 
         if len(mandatory_conditions) > 0:
             query.add_condition_to_ast(combine_and_conditions(mandatory_conditions))
